+++ package.json
@@ -10,7 +10,6 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.1.3",
-    "formik": "^2.2.9",
     "history": "^5.3.0",
     "js-cookie": "^3.0.1",
     "jwt-decode": "^3.1.2",
@@ -20,7 +19,6 @@
     "react-phone-input-2": "^2.15.1",
     "react-router-dom": "^6.4.2",
     "react-scripts": "5.0.1",
-    "validator": "^13.7.0",
     "web-vitals": "^2.1.4",
     "workbox-background-sync": "^6.5.4",
     "workbox-broadcast-update": "^6.5.4",
