+++ src/routers.js
@@ -6,7 +6,7 @@ import { NonAuthRoutes, AuthRoutes } from "./url";
 import { ReactComponent as LoadingIcon } from "./assets/svg/loading-icon.svg";
 
 const LandingPage = React.lazy(() => import("./pages/landingPage"));
-const SignUp = React.lazy(() => import("./pages/signUp"));
+const SignUp = React.lazy(() => import("./pages/signUp/signUp"));
 const Login = React.lazy(() => import("./pages/login"));
 const VerifyEmail = React.lazy(() => import("./pages/verifyEmail"));
 const EnterEmailVerificationCode = React.lazy(() =>
@@ -36,37 +36,37 @@ function Routers() {
         }
       >
         <Routes>
-          <Route path={NonAuthRoutes.landingPage} element={<LandingPage />} />
-          <Route path={NonAuthRoutes.signUp} element={<SignUp />} />
-          <Route path={NonAuthRoutes.login} element={<Login />} />
+          <Route path={NonAuthRoutes.landingPage} element={<LandingPage />} />{" "}
+          <Route path={NonAuthRoutes.signUp} element={<SignUp />} />{" "}
+          <Route path={NonAuthRoutes.login} element={<Login />} />{" "}
           <Route
             path={NonAuthRoutes.forgotPassword}
             element={<ForgotPassword />}
-          />
+          />{" "}
           <Route
             path={NonAuthRoutes.resetPassword}
             element={<ResetPassword />}
-          />
+          />{" "}
           <Route
             path={NonAuthRoutes.alertForgotPassword}
             element={<AlertForgotPassword />}
-          />
+          />{" "}
           <Route
             path={NonAuthRoutes.alertResetPassword}
             element={<AlertResetPassword />}
-          />
-          <Route path={NonAuthRoutes.verifyEmail} element={<VerifyEmail />} />
+          />{" "}
+          <Route path={NonAuthRoutes.verifyEmail} element={<VerifyEmail />} />{" "}
           <Route
             path={NonAuthRoutes.enterEmailVerificationCode}
             element={<EnterEmailVerificationCode />}
-          />
+          />{" "}
           <Route
             path={NonAuthRoutes.enterPasswordResetCode}
             element={<EnterPasswordResetCode />}
-          />
-          <Route path={AuthRoutes.dashboard} element={<Dashboard />} />
-        </Routes>
-      </Suspense>
+          />{" "}
+          <Route path={AuthRoutes.dashboard} element={<Dashboard />} />{" "}
+        </Routes>{" "}
+      </Suspense>{" "}
     </div>
   );
 }
